Crashed Exp

+++ src/main/java/generator/BasicRandomGenerator.java
@@ -4,7 +4,7 @@ import java.util.Random;
 
 public class BasicRandomGenerator implements IRandomGenerator{
     @Override
-    public int generate() {
+    public double generate() {
         Random rand = new Random();
         int randomNum = rand.nextInt(5000) + 1;
         return randomNum;
