+++ src/main/java/generator/IRandomGenerator.java
@@ -1,5 +1,5 @@
 package generator;
 
 public interface IRandomGenerator {
-    int generate();
+    double generate();
 }
