+++ src/main/java/generator/WeibullRandomGenerator.java
@@ -4,11 +4,11 @@ import org.apache.commons.math3.distribution.WeibullDistribution;
 
 public class WeibullRandomGenerator implements IRandomGenerator{
     @Override
-    public int generate() {
+    public double generate() {
         double shape = 15; // shape parameter of the Weibull distribution
         double scale = 5000; // scale parameter of the Weibull distribution
         WeibullDistribution weibull = new WeibullDistribution(shape, scale);
         double randomNum = weibull.sample();
-        return (int) randomNum;
+        return randomNum;
     }
 }
